refactor(client): 将客户端流体模型注册迁移到独立类
- 从 ModFluids 类中移除流体模型注册相关代码
- 新建 ClientModFluidModels 类专责客户端流体模型注册
- 保持流体模型的材质和色调逻辑不变
- 优化代码结构，增强职责单一性和可维护性
